requirements.txt
pytest~=8.4.1
selene==2.0.0rc9
allure-pytest
requests~=2.32.4
python-dotenv~=1.0.1
utils~=1.0.2
allure-python-commons~=2.14.3